pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strands-rag-pipeline"
version = "1.0.0"
description = "A multimodal agentic RAG pipeline built with AWS Strands Agents SDK and CDK infrastructure"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Mohammed Firdous"}
]
keywords = ["rag", "ai", "strands", "aws", "cdk", "lambda", "agents"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "strands-agents",
    "strands-agents-tools",
    "aws-cdk-lib>=2.0.0",
    "constructs>=10.0.0",
    "boto3",
    "langchain",
    "langchain-community",
    "chromadb",
    "pypdf2",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "pylint>=2.17.5",
    "mypy>=1.4.1",
    "pre-commit>=3.3.3",
    "ipython>=8.14.0",
    "ipdb>=0.13.13",
]

[project.urls]
Homepage = "https://github.com/mohammedfirdouss/strands-rag-pipeline"
Repository = "https://github.com/mohammedfirdouss/strands-rag-pipeline"
Documentation = "https://github.com/mohammedfirdouss/strands-rag-pipeline/blob/main/README.md"

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | cdk.out
)/
'''

[tool.pylint.messages_control]
max-line-length = 120
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "R0913",  # too-many-arguments
    "R0914",  # too-many-locals
]

[tool.pylint.format]
max-line-length = 120

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=agents",
    "--cov=lambda",
    "--cov=infrastructure",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = [
    "strands.*",
    "strands_tools.*",
    "langchain.*",
    "chromadb.*",
]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = [".venv", "venv", "tests", "cdk.out"]
skips = ["B101"]  # Skip assert_used test
